Module rsa_cortex_m4::arithmetic

Modular arithmetic (for numbers that are either Convenient or word-sized powers-of-two).

For Unsigned, we implement operations "$\text{mod } 2^{32(D + E)}$", that is, dropping all carries and borrows.

For Modular, we use incompletely reduced representations internally (which can be implemented on a word-level), offering a complete reduction for external use (which needs to be implemented on a bit-level).

We do indeed have use for the modular interpretation of Unsigned, for instance, to calculate $65537^{-1} \text{ mod }(p - 1)$.

Structs

Modular
Modular integer, corresponds to the residue class "modulo modulus".

Montgomery
Montgomery representation of $[x]_{n} := x\text{ }(\text{mod }n)$, as $[x \cdot 2^{-32L}]_n$.

Type Definitions

ShortModular

ShortMontgomery