Struct rsa_cortex_m4::numbers::Convenient

#[repr(transparent)]pub struct Convenient<const D: usize, const E: usize>(_);

Unsigned numbers with both their top and bottom bits set – highly convenient for modular arithmetic!

In particular, they are odd. But also, $n \ge 2^{m - 1}$, with strict inequality by oddness.

As described in Incomplete reduction in modular arithmetic (2002), it is not necessary to reduce fully modulo n while calculating modular arithmetic. Instead, we can reduce modulo $2^m$, and only "fully" reduce when so desired.

Their arguments apply to non-prime moduli also, and the "convenient" ones have the properties, in their terminology, that $I = 1$ and $J = 2$, hence $F = 2^m - p$ and $G = 2p - 2^m$. Moreover (!!!), the last case/reduction in their modular addition / subtraction never occurs.

E.g., in addition, we have $F = 2^m - n < 2^m - 2^{m - 1} = 2^{m-1}$, and so $T := S + F < (2^m - 2) + 2^{m - 1} < 2^{m - 1}$.