Struct rsa_cortex_m4::numbers::Unsigned

#[repr(C)]pub struct Unsigned<const D: usize, const E: usize> { /* fields omitted */ }

Multi-precision unsigned integer with at most $D + E$ digits (places) – two Limbs.

Workaround type for limitations of const generics on stable; the interesting cases are:

Short, where $E = 0$, and

Long, where $D = E$.

The former is used for RSA prime pairs $(P, Q)$, the latter for RSA public keys $N = PQ$.

Mnemonics: D for digits, E for "extra" digits.

Implementations

impl<const D: usize, const E: usize> Unsigned<D, E>

pub fn checked_add(&self, summand: &Self) -> Option<Self>

pub fn wrapping_add_assign(&mut self, summand: &Self)

pub fn wrapping_add(&self, summand: &Self) -> Self

impl<const D: usize, const E: usize> Unsigned<D, E>

pub fn wrapping_neg(&self) -> Self

pub fn checked_sub(&self, subtrahend: &Self) -> Option<Self>

pub fn wrapping_sub_assign(&mut self, subtrahend: &Self)

pub fn wrapping_sub(&self, subtrahend: &Self) -> Self

impl<const D: usize, const E: usize> Unsigned<D, E>

pub fn wrapping_mul(&self, factor: &Self) -> Self

impl<const D: usize, const E: usize> Unsigned<D, E>

pub fn wrapping_inv(&self) -> Result<Self>

The wrapping inverse, i.e., the exact inverse w.r.t wrapping multiplication.

Exists if and only if the number is odd.

This uses $\mathcal{O}(\log n)$ loops in Self::BITS, very efficient (!)

Source: Fig. 1 from GCD-Free Algorithms for Computing Modular Inverses (2003)

Note that this source is highly confusing! What they mean to say is to iterate $y \leftarrow y(2 - ey)$ in $\mathbb{Z}/2^{|f|}$, where the output is an inverse of $e$ modulo $2^{2i}$. In other words, the $\text{mod }2^i$ is a typo, and should be $\text{mod }2^{|f|}$.

cf. also Crypto StackExchange.

impl<const D: usize, const E: usize> Unsigned<D, E>

Reduction of unsigned integers

pub fn modulo<const F: usize, const G: usize>(
self,
n: &Convenient<F, G>
) -> Modular<'_, F, G>

The associated residue class modulo n.

Note that storage requirements of the residue class are the same as the modulus (+ reference to it), not the original integer.

This uses incomplete reduction (Self::partially_reduce) for efficiency.

pub fn partially_reduce<const F: usize, const G: usize>(&self) -> Unsigned<F, G>

A noncanonical representative of the associated residue class modulo n.

The "incomplete reduction" modulo $w^{D + E}$, where $w$ is the digit size $2^{\text{BITS}}$, i.e., the word size of the machine.

Cf. Modular.

pub fn reduce<const F: usize, const G: usize>(
&self,
n: &Unsigned<F, G>
) -> Unsigned<F, G>

The canonical (completely) reduced representative of the associated residue class modulo $n$.

Cf. Modular.

impl<const D: usize, const E: usize> Unsigned<D, E>

pub fn to_be_bytes(&self) -> BigEndian<D, E>

TODO: consider into_be_bytes, reusing the buffer.

i.e.

swap words + endianness on self.0

return BigEndian(self.0)

impl<const D: usize, const E: usize> Unsigned<D, E>

Trait methods as inherent methods, for convenience.

pub fn from_slice(slice: &[Digit]) -> Self

pub fn try_from_slice(slice: &[Digit]) -> Result<Self>

pub fn leading_digit(&self) -> Option<Digit>

pub fn try_into_unsigned<const M: usize, const N: usize>(
&self
) -> Result<Unsigned<M, N>>

pub fn into_unsigned<const M: usize, const N: usize>(&self) -> Unsigned<M, N>

pub fn one() -> Self

pub fn zero() -> Self