Trait rsa_cortex_m4::numbers::Number

pub unsafe trait Number: Deref<Target = [Digit]> + Clone + Debug + Default {
    const CAPACITY: usize;
    fn significant_digits(&self) -> &[Digit] { ... }

    fn leading_digit(&self) -> Option<Digit> { ... }

    fn try_into_unsigned<const D: usize, const E: usize>(
        &self
    ) -> Result<Unsigned<D, E>> { ... }

    fn to_unsigned<const D: usize, const E: usize>(&self) -> Unsigned<D, E> { ... }

    fn zero() -> Self { ... }

    fn is_zero(&self) -> bool { ... }

    fn is_one(&self) -> bool { ... }

    fn is_digit(&self) -> bool { ... }

    fn is_odd(&self) -> bool { ... }

    fn cmp(&self, other: &impl Number) -> Ordering { ... }

    fn eq(&self, other: &impl Number) -> bool { ... }

    fn deref(&self) -> &[Digit] { ... }

    fn deref_mut(&mut self) -> &mut Self::Target { ... }
}

Something similar to a Vec<u32>, without allocations.

The dereferenced slice is treated as little-endian digits of a big unsigned integer; this slice must be of length Self::CAPACITY.

There is no need to "extend" the allocation as in, say, heapless. Simply write to the desired index / slice (via DerefMut).

In a previous version of this trait, only the significant digits (up until the leading digit) were dereferenced. To meet constant-time requirements, this was changed.

Current implementations are (with const generic usize parameters):

Unsigned<D, E>

Array<D, E, L>

Of actual interest are Long (=Unsigned<D, D>) and Short (=Unsigned<D, 0>) numbers, where "Short" is tongue-in-cheek.

A lot of this dance could be skipped if only sums of const generic usizes were considered const (which they are not in Rust 1.51's min_const_generics.

All we really want is to have two "Short" primes $P, Q$, and their "Long" product $N = PQ$.

Associated Constants

const CAPACITY: usize

Loading content...
Provided methods

fn significant_digits(&self) -> &[Digit]

The significant digits of the number (little-endian).

fn leading_digit(&self) -> Option<Digit>

The last non-zero digit of the number.

fn try_into_unsigned<const D: usize, const E: usize>(
&self
) -> Result<Unsigned<D, E>>

Embed in number with D + E digits, if possible.

Not expressable as TryInto, as it would clash with blanket implementations, e.g. for Unsigned with D = X.

fn to_unsigned<const D: usize, const E: usize>(&self) -> Unsigned<D, E>

Panics iff Self::try_into_unsigned fails.

Internal use of this embedding of abstract Numbers in Unsigneds never actually panics, bar implementation errors.

fn zero() -> Self

fn is_zero(&self) -> bool

fn is_one(&self) -> bool

fn is_digit(&self) -> bool

fn is_odd(&self) -> bool

fn cmp(&self, other: &impl Number) -> Ordering

This is little endian ordering, as opposed to the default ordering on arrays and slices!

In other words, we start comparing at the leading digits.

fn eq(&self, other: &impl Number) -> bool

fn deref(&self) -> &[Digit]

fn deref_mut(&mut self) -> &mut Self::Target

Loading content...