Struct rsa_cortex_m4::F4

pub struct F4 {}

The fourth Fermat prime, $2^{16} + 1$ (used as public exponent $e$).

This library only implements RSA keys with public exponent e = 65537 = 0x10001 = u16::MAX + 2.

An example recommendation to do so is RFC 4871: https://www.ietf.org/rfc/rfc4871.txt, more generally, there seems no need to have too many knobs to turn.

Implementations

impl F4

pub const DIGIT: Digit

pub const SHORT: Short<1>

NOTE!!! F4 is not convenient, it's only a 17-bit number, whereas Convenient<1, 0> would mean either 31-bit or 63-bit.

pub const PRIME: Prime<1, 0>

pub fn minus_one() -> Short<1>

pub fn prime() -> Prime<1, 0>

pub fn wrapping_inv<const D: usize, const E: usize>() -> Unsigned<D, E>

pub fn inv_mod<const D: usize>(p: &Unsigned<D, 0>) -> Odd<D, 0>

The inverse of $F4$ modulo other primes is used in RSA, and deserves an optimized implementation.

This is Arazi's lemma.

TODO: Prove (disprove?) that calculating the inverse of p_mod_e via PrimeModular's inversion is valid, as F4 is not actually "convenient".

N.B.: We do have a test that the inverse is correct (see test inv_mod_e).