Struct rsa_cortex_m4::PrivateKey

pub struct PrivateKey<const D: usize> { /* fields omitted */ }

RSA private key.

Fundamentally, this consists of two different primes p and q, which should both have bit length 8L (i.e., L4 bytes). Additionally, d_p = e^{-1} (mod p) and d_q = e^{-1} (mod q) are stored.

It's quite sad, but we can't enforce the bound L2 = 2*L.

Implementations

impl<const D: usize> PrivateKey<D>

pub fn decryption_primitive(&self, ciphertext: &[u8]) -> Result<Long<D>>

pub fn classic_decryption_primitive(&self, ciphertext: &[u8]) -> Result<Long<D>>

RSADP

pub fn inversion_free_decryption(&self, ciphertext: &[u8]) -> Result<Long<D>>

This saves calculating and storing the inverse $q^{-1}\text{ mod }p$.

Source

pub fn blinded_inversion_free_decryption(
&self,
rng: impl CryptoRng + RngCore,
ciphertext: &[u8]
) -> Result<Long<D>>

As also explained in Improvements..., it is easy to blind the factors.

pub fn signature_primitive(&self, msg: &[u8]) -> Result<Long<D>>

RSASP1

pub fn sign<P, R>(&self, msg: &[u8], _padding: P, rng: R) -> Result<Long<D>> where
P: SignaturePadding<D>,
R: CryptoRng + RngCore,

Sign a message with respect to a padding method.

pub fn decrypt<P>(&self, ciphertext: &[u8], _: P) -> Result<Unpadded<D>> where
P: EncryptionPadding<D>,

Decrypt a ciphertext with respect to a padding method.

impl<const L: usize> PrivateKey<L>

pub fn new(_rng: impl RngCore + CryptoRng) -> Result<Self>