Trait rsa_cortex_m4::numbers::Number

pub unsafe trait Number: Deref<Target = [Digit]> {
    const CAPACITY: usize;

    fn len(&self) -> usize;

    fn capacity() -> usize { ... }

    fn number(&self) -> &[Digit] { ... }

    fn padded_number(&self) -> &[Digit] { ... }

    fn leading_digit(&self) -> Option<Digit> { ... }

    fn try_into_unsigned<const D: usize, const E: usize>(
        &self
    ) -> Result<Unsigned<D, E>> { ... }

    fn into_unsigned<const D: usize, const E: usize>(&self) -> Unsigned<D, E> { ... }
}

Something similar to a Vec<u32>, without allocations.

Implementation must ensure:

self.len() == self.number().len()

self.capacity() == self.number_mut().len()

Deref coincides with self.number()

DerefMut coincides with self.padded_number_mut()

It may or may not be the case that having Deref/DerefMut of different length is too cute (cf. remarks on NumberMut::invalidate_len), but it's terribly convenient:

There is no need to "extend" the allocation, simply write to the desired index / slice (via DerefMut).

For efficiency, our implementations track (cache) len.

The highest non-zero term determines the leading digit and the length.

Current implementations are (with const generic usize parameters):

Unsigned<D, E>

Array<D, E, L>

Of actual interest are Long (=Unsigned<D,D>) and Short (=Unsigned<D,)>) numbers, where "Short" is tongue-in-cheek.

A lot of this dance could be skipped if only sums of const generic usizes were considered const (which they are not in Rust 1.51's min_const_generics.

All we really want is to have two "Short" primes $P, Q$, and their "Long" product $N = PQ$.

Associated Constants

const CAPACITY: usize

Loading content...
Required methods

fn len(&self) -> usize

The length of the number in terms of relevant digits.

Example: [0, 1, 0, 2, 0, 0, 0, 0] has length 4.

Loading content...
Provided methods

fn capacity() -> usize

fn number(&self) -> &[Digit]

Default implementation assumes "data slice" starts at object address.

fn padded_number(&self) -> &[Digit]

Default implementation assumes "data slice" starts at object address.

fn leading_digit(&self) -> Option<Digit>

fn try_into_unsigned<const D: usize, const E: usize>(
&self
) -> Result<Unsigned<D, E>>

Embed in number with D digits, if possible.

Fails: iff self.len() > D + E.

Not expressable as TryInto, as it would clash with blanket implementations, e.g. for Unsigned with D = X.

fn into_unsigned<const D: usize, const E: usize>(&self) -> Unsigned<D, E>

Panics iff Self::try_into_unsigned fails.

Internal use of this embedding of abstract Numbers in Unsigneds never actually panics, bar implementation errors.

Loading content...